requirements.txt
torch
torchvision
Pillow
accelerate
pdf2image
cmake
pkgconfig
poppler-utils
layoutparser
opencv-python
layoutparser[ocr]
pytesseract
pymupdf
tabula-py
img2table
llama-index
llama-index-core
llama-index-embeddings-huggingface
llama-index-postprocessor-flag-embedding-reranker
llama-index-vector-stores-pinecone
llama-index-vector-stores-chroma
llama-index-vector-stores-astra_db
llama-index-llms-llama_cpp
pinecone-client
llama-parse
bitsandbytes
llama-cpp-python
transformers
scikit-learn
chromadb